Fix stale closures of callback options in inputs.text et. al.
The implementation of useFormState is problematic because it
caches callback functions passed to inputs.text et. al. This
causes functions like onChange, validate, and onBlur to run with
stale closures, namely, the closure with which they are
initialised.

This commit fixes the issue by removing the `callback` cache from
the implementation. Tests that depended on this buggy behavior
were removed as they are unsupportable features given the current
implementation.